polish seed code
