+++ requirements.txt
@@ -12,3 +12,4 @@ cython
 lxml
 premailer
 openpyxl
+fasttext==0.9.1
